TimelineEntry: moves touch tracking to GestureDetector
In order to better support multiple DPIs as well as to be able to detect
taps and such, this gets rid of the custom finger tracking in order to
detect fling / scroll gestures. Unfortunately, this also gets rid of
pinch to zoom in/out, but it was never working flawlessly anyhow.